requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
cloudinary==1.40.0
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
Django==5.0.4
django-allauth==0.62.1
django-cors-headers==4.3.1
django-stubs==5.0.2
django-stubs-ext==5.0.2
djangorestframework==3.15.1
Faker==25.0.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
mypy==1.10.0
mypy-extensions==1.0.0
oauthlib==3.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
requests==2.31.0
requests-oauthlib==2.0.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
toposort==1.10
Twisted==24.3.0
txaio==23.1.1
types-PyYAML==6.0.12.20240311
typing_extensions==4.11.0
urllib3==2.2.1
zope.interface==6.3
